Arduino UNO R4 MINIMをDAPLINKでプログラミング

+++ pac/src/main.rs
@@ -197,7 +197,7 @@ fn clock_init_hoco48(p: &pac::Peripherals) {
     // HOCOCR2レジスタのアドレス: 0x4001_e037
     // HOCO48MHz指定: 0b0010_0000
     unsafe {
-        core::ptr::write_volatile(0x4001_e037 as *mut u32, 0b0010_0000);
+        core::ptr::write_volatile(0x4001_e037 as *mut u8, 0b0010_0000);
     }
     // 高速オンチップオシレータ(HOCO)クロック動作
     p.SYSTEM.hococr().write(|w| w.hcstp()._0());
